Constructing the efficient frontier by using Python and Excel for five stocks which are selected from a capital market
The most important area in finance and investment analysis is the relationship between the expected rate of return and the risk of asset.  In this project, we have examined this relationship and also how that can help the investors to make better investment decisions i.e. high level of return for the lowest level of risk.